QAP-8 | Refactor for sequelize migrate, add privacy_policy column.
