--- web/src/database/definitions/Test.js
@@ -1,46 +1,0 @@
-'use strict';
-
-const DataTypes = require('sequelize');
-
-// @todo: Migrations.
-// @see: http://docs.sequelizejs.com/manual/tutorial/migrations.html
-module.exports = {
-    attributes: {
-        id: {
-            type: DataTypes.INTEGER,
-            allowNull: false,
-            primaryKey: true,
-            autoIncrement: true
-        },
-        reference_url: {
-            type: DataTypes.TEXT,
-            allowNull: false
-        },
-        test_url: {
-            type: DataTypes.TEXT,
-            allowNull: false
-        },
-        email: {
-            type: DataTypes.STRING,
-            allowNull: false
-        },
-        uuid: {
-            type: DataTypes.UUID,
-            defaultValue: DataTypes.UUIDV4,
-            allowNull: false,
-            unique: true
-        },
-        newsletter: {
-            type: DataTypes.BOOLEAN,
-            allowNull: false
-        },
-        createdAt: {
-            type: DataTypes.DATE,
-            allowNull: false,
-            defaultValue: DataTypes.NOW
-        }
-    },
-    options: {
-        timestamps: false
-    }
-};
