+++ web/src/route/api-v1.js
@@ -79,11 +79,11 @@ apiRouter.post('/test/add', asyncHandler(async function (req, res) {
     /**
      * @type {Sequelize.Model}
      */
-    const Test = db.models.Test;
+    const Tests = db.models.Tests;
     let existingTests = 0;
 
     try {
-        existingTests = await Test.count({
+        existingTests = await Tests.count({
             where: { email: data.email }
         });
     }
@@ -108,7 +108,7 @@ apiRouter.post('/test/add', asyncHandler(async function (req, res) {
 
     try {
         // The validation only allows the required fields, so this should be OK.
-        newTest = await Test.create(data);
+        newTest = await Tests.create(data);
     }
     catch (error) {
         return res.status(500).json({
@@ -150,11 +150,11 @@ apiRouter.post('/test/get', asyncHandler(async function (req, res) {
     /**
      * @type {Sequelize.Model}
      */
-    const Test = db.models.Test;
+    const Tests = db.models.Tests;
 
     let existingTest = undefined;
     try {
-        existingTest = await Test.findOne({ where: { uuid: uuid }});
+        existingTest = await Tests.findOne({ where: { uuid: uuid }});
     }
     catch (error) {
         return res.status(500).json({
